+++ R/error-checks.R
@@ -71,13 +71,18 @@ isOfType <- function(object, type, nullAllowed = FALSE) {
 #' isIncluded(list("x", 1), list("a", "b", "x")) # FALSE
 #' @export
 isIncluded <- function(values, parentValues) {
+  values <- c(values)
+  hasObject <- any(mapply(function(x) {
+    is.object(x)
+  }, values))
+
+  if (hasObject) {
+    stop("Only vectors of base object types are allowed.", call. = FALSE)
+  }
+
   if (is.null(values) || length(values) == 0) {
     return(FALSE)
   }
-
-  # make sure they are vectors
-  values <- .toVector(values)
-  parentValues <- .toVector(parentValues)
 
   as.logical(min(values %in% parentValues))
 }
@@ -157,18 +162,6 @@ hasUniqueValues <- function(values, na.rm = TRUE) {
   }
 
   return(!any(duplicated(values)))
-}
-
-
-# utilities ---------------------------------------------
-
-#' @keywords internal
-.toVector <- function(x) {
-  if (!is.vector(x)) {
-    x <- c(x)
-  }
-
-  return(x)
 }
 
 #' @keywords internal