+++ NEWS.md
@@ -1,5 +1,9 @@
 # ospsuite.utils 1.0.0.9000
 
+MAJOR CHANGES
+
+* `isIncluded()` now only accepts base types as valid inputs.
+
 MINOR CHANGES
 
 * Improvements to documentation.